refactor: slit reporter tests to match code
